+++ R/elx_make_query.R
@@ -17,6 +17,8 @@
 #' @param include_lbs If `TRUE`, results include legal bases of legislation
 #' @param include_force If `TRUE`, results include whether legislation is in force
 #' @param include_eurovoc If `TRUE`, results include EuroVoc descriptors of subject matter
+#' @param include_author If `TRUE`, results include document author(s)
+#' @param include_citations If `TRUE`, results include citations (CELEX-labelled)
 #' @param order Order results by ids
 #' @param limit Limit the number of results, for testing purposes mainly
 #' @return
@@ -32,9 +34,7 @@ elx_make_query <- function(resource_type = c("directive","regulation","decision"
                            manual_type = "", include_corrigenda = FALSE, include_celex = TRUE, include_lbs = FALSE,
                            include_date = FALSE, include_date_force = FALSE, include_date_endvalid = FALSE,
                            include_date_transpos = FALSE, include_force = FALSE, include_eurovoc = FALSE,
-                           order = FALSE, limit = NULL){
-
-  # for the moment, 'councilvotes' is not an operational resource type
+                           include_author = FALSE, include_citations = FALSE, order = FALSE, limit = NULL){
 
   if (!resource_type %in% c("directive","regulation","decision","recommendation","intagr","caselaw","manual","proposal","national_impl")) stop("'resource_type' must be defined")
 
@@ -108,6 +108,18 @@ elx_make_query <- function(resource_type = c("directive","regulation","decision"
   if (include_eurovoc == TRUE){
 
     query <- paste(query, "?eurovoc", sep = " ")
+
+  }
+
+  if (include_author == TRUE){
+
+    query <- paste(query, "?author", sep = " ")
+
+  }
+
+  if (include_citations == TRUE){
+
+    query <- paste(query, "?citation ?citationcelex", sep = " ")
 
   }
 
@@ -235,35 +247,33 @@ elx_make_query <- function(resource_type = c("directive","regulation","decision"
     query <- paste(query,"\n FILTER not exists{?work cdm:work_has_resource-type <http://publications.europa.eu/resource/authority/resource-type/CORRIGENDUM>}", sep = " ")
   }
 
-  #query <- paste(query, "?work cdm:work_id_document ?doc_id.")
-
   if (include_celex == TRUE){
 
-    query <- paste(query, "?work cdm:resource_legal_id_celex ?celex.")
+    query <- paste(query, "OPTIONAL{?work cdm:resource_legal_id_celex ?celex.}")
 
   }
 
   if (include_date == TRUE){
 
-    query <- paste(query, "?work cdm:work_date_document ?date.")
+    query <- paste(query, "OPTIONAL{?work cdm:work_date_document ?date.}")
 
   }
 
   if (include_date_force == TRUE){
 
-    query <- paste(query, "?work cdm:resource_legal_date_entry-into-force ?dateforce.")
+    query <- paste(query, "OPTIONAL{?work cdm:resource_legal_date_entry-into-force ?dateforce.}")
 
   }
 
   if (include_date_endvalid == TRUE){
 
-    query <- paste(query, "?work cdm:resource_legal_date_end-of-validity ?dateendvalid.")
+    query <- paste(query, "OPTIONAL{?work cdm:resource_legal_date_end-of-validity ?dateendvalid.}")
 
   }
 
   if (include_date_transpos == TRUE){
 
-    query <- paste(query, "?work cdm:directive_date_transposition ?datetranspos.")
+    query <- paste(query, "OPTIONAL{?work cdm:directive_date_transposition ?datetranspos.}")
 
   }
 
@@ -281,14 +291,26 @@ elx_make_query <- function(resource_type = c("directive","regulation","decision"
 
   if (include_force == TRUE){
 
-    query <- paste(query, "?work cdm:resource_legal_in-force ?force.")
+    query <- paste(query, "OPTIONAL{?work cdm:resource_legal_in-force ?force.}")
 
   }
 
   if (include_eurovoc == TRUE){
 
-    query <- paste(query, '?work cdm:work_is_about_concept_eurovoc ?eurovoc. graph ?gs
-    { ?eurovoc skos:prefLabel ?subjectLabel filter (lang(?subjectLabel)="en") }.')
+    query <- paste(query, 'OPTIONAL{?work cdm:work_is_about_concept_eurovoc ?eurovoc. graph ?gs
+    { ?eurovoc skos:prefLabel ?subjectLabel filter (lang(?subjectLabel)="en") }.}')
+
+  }
+
+  if (include_author == TRUE){
+
+    query <- paste(query, "OPTIONAL{?work cdm:work_created_by_agent ?author.}")
+
+  }
+
+  if (include_citations == TRUE){
+
+    query <- paste(query, "OPTIONAL{?work cdm:work_cites_work ?citation. ?citation cdm:resource_legal_id_celex ?citationcelex.}")
 
   }
 
@@ -303,3 +325,4 @@ elx_make_query <- function(resource_type = c("directive","regulation","decision"
   return(query)
 
 }
+